+++ packages/react/src/patterns/F0Form/fields/types.ts
@@ -183,6 +183,7 @@ export type FieldType =
   | "daterange"
   | "period"
   | "phone"
+  | "location"
   | "richtext"
   | "file"
   | "cardSelect"
@@ -223,6 +224,11 @@ export type {
   F0PhoneInputValue,
 } from "./phone/types"
 export type {
+  F0LocationConfig,
+  F0LocationField,
+  F0LocationInputValue,
+} from "./location/types"
+export type {
   F0RichTextConfig,
   F0RichTextField,
   RichTextValue,
@@ -262,6 +268,7 @@ import type { F0CustomField } from "./custom/types"
 import type { F0DurationField } from "./duration/types"
 import type { F0EntitiesListField } from "./entitiesList/types"
 import type { F0FileField } from "./file/types"
+import type { F0LocationField } from "./location/types"
 import type { F0PhoneField } from "./phone/types"
 import type { F0RichTextField } from "./richtext/types"
 import type { F0SwitchField } from "./switch/types"
@@ -285,6 +292,7 @@ export type F0Field =
   | F0DateRangeField
   | F0PeriodField
   | F0PhoneField
+  | F0LocationField
   | F0RichTextField
   | F0FileField
   | F0CardSelectField